+++ portal-impl/src/com/liferay/portlet/internal/PortletRequestDispatcherImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet;
+package com.liferay.portlet.internal;
 
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringPool;
@@ -31,6 +31,11 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.servlet.DynamicServletRequestUtil;
 import com.liferay.portal.struts.StrutsURLEncoder;
+import com.liferay.portlet.MimeResponseImpl;
+import com.liferay.portlet.PortletRequestImpl;
+import com.liferay.portlet.PortletResponseImpl;
+import com.liferay.portlet.PortletServletRequest;
+import com.liferay.portlet.PortletServletResponse;
 
 import java.io.IOException;
