+++ portal-impl/test/unit/com/liferay/portlet/PortletRequestDispatcherImplTest.java
@@ -25,6 +25,7 @@ import com.liferay.portal.model.impl.PortletAppImpl;
 import com.liferay.portal.model.impl.PortletImpl;
 import com.liferay.portal.util.PortalImpl;
 import com.liferay.portlet.internal.PortletContextImpl;
+import com.liferay.portlet.internal.PortletRequestDispatcherImpl;
 import com.liferay.registry.BasicRegistryImpl;
 import com.liferay.registry.RegistryUtil;
